+++ src/main/java/dev/codehouse/backend/admin/service/AdminProblemService.java
@@ -2,7 +2,7 @@ package dev.codehouse.backend.admin.service;
 
 import dev.codehouse.backend.admin.dto.ProblemRequest;
 import dev.codehouse.backend.admin.dto.ProblemResponse;
-import dev.codehouse.backend.admin.entity.Problem;
+import dev.codehouse.backend.problem.entity.Problem;
 import dev.codehouse.backend.admin.repository.ProblemRepository;
 import dev.codehouse.backend.global.exception.AdminException;
 import dev.codehouse.backend.global.response.ResponseCode;
